fix: added interitedProps handler to main renderer
- added `inheritedProps` handler to the enable the ability to pass props
futher down the component tree. While most will consider this an
anti-pattern (it is, no argument there), its use in this project is
limited (but nessessary) to currently the core/column and core/columns
blocks in order to pass down column specific props. This is the only
situation so far where two blocks need to share props, and given the
dynamic nature of potential layouts changes, it is not feasible to use
other methods like context or data stores.