refactor: changes all across the codebase
- refactor: blockColumns passes the number of columns down (via props)
to blockColumn to set the column span.

- refactor: blockGroup completely rebuild to handle the 4 different
layout options (no-layout, flex col, flex row, and grid) as well as
gracefully handling layout options for no-layout blocks.

- refactor: moved normaliseClassNames to a new file @/lib/styler.js
and renamed it to styler. It's much nice to work with now, and can be
extended by adding more functions to the styler.js file.

- refactor: removed normaliseClassNamesFromAttributes from @/lib/utils
and created a new function in @/lib/attributes.js called
processAttributesToClassNames, which processes camelCase attributes
to Tailwind kebab-case class names. This new function has been written
to not required any dependencies as it is also used in the node build
script scripts/build-safelist.js.

- refactor: removed withConditionalInnerWrapper from @/lib/utils and
all blocks that used it as that has been removed from the server
side codebase.

- refactor: various codebase improvements including:
-- removed unused imports, variables, and functions
-- inlined and removed single-use functions
-- removed unused files